+++ doc/source/changelog.rst
@@ -9,6 +9,51 @@ This document contains the release notes for the project.
 
 .. towncrier release notes start
 
+`0.4.2 <https://github.com/ansys/pyconceptev/releases/tag/v0.4.2>`_ - 2024-09-13
+================================================================================
+
+Changed
+^^^^^^^
+
+- Testing, do not merge `#114 <https://github.com/ansys/pyconceptev/pull/114>`_
+
+
+Fixed
+^^^^^
+
+- added a product id getter `#111 <https://github.com/ansys/pyconceptev/pull/111>`_
+
+
+Miscellaneous
+^^^^^^^^^^^^^
+
+- CHORE: Revert changelog changes after release failure `#112 <https://github.com/ansys/pyconceptev/pull/112>`_
+
+`0.4.1 <https://github.com/ansys/pyconceptev/releases/tag/v0.4.1>`_ - 2024-09-12
+================================================================================
+
+Changed
+^^^^^^^
+
+- chore: update CHANGELOG for v0.4.0 `#106 <https://github.com/ansys/pyconceptev/pull/106>`_
+- CHORE: Bump development version to 0.5.dev0 `#107 <https://github.com/ansys/pyconceptev/pull/107>`_
+- BUILD: Update dependencies range `#109 <https://github.com/ansys/pyconceptev/pull/109>`_
+
+`0.4 <https://github.com/ansys/pyconceptev/releases/tag/v0.4>`_ - 2024-09-04
+============================================================================
+
+Added
+^^^^^
+
+- Added progress monitor and tests `#105 <https://github.com/ansys/pyconceptev/pull/105>`_
+
+
+Changed
+^^^^^^^
+
+- chore: update CHANGELOG for v0.3.0 `#102 <https://github.com/ansys/pyconceptev/pull/102>`_
+- CHORE: Bump dev version into 0.4.dev0 `#103 <https://github.com/ansys/pyconceptev/pull/103>`_
+
 `0.3.0 <https://github.com/ansys/pyconceptev/releases/tag/v0.3.0>`_ - 2024-08-22
 ================================================================================